ci(win): only remove choco's cpack.exe if it exists

+++ ci/build.ps1
@@ -171,7 +171,9 @@ if (-not $NoTests) {
 }
 
 # Ensure choco's cpack is not in PATH otherwise, it conflicts with CMake's
-Remove-Item -Path $env:ChocolateyInstall\bin\cpack.exe -Force
+if (Test-Path -Path $env:ChocolateyInstall\bin\cpack.exe) {
+  Remove-Item -Path $env:ChocolateyInstall\bin\cpack.exe -Force
+}
 
 # Build artifacts
 cpack -G ZIP -C RelWithDebInfo
